fix(items): fold changeOrder enums into items.models to satisfy module-shape
The PLM foundation added a separate changeOrder.models.ts, which the
module-shape conformance gate rejects (one <module>.models.ts per module).
Move the enum mirrors into items.models.ts and drop the barrel re-export.
